colour_following_mouse_point.py
this is python simple program which is made by using opencv numpy and windows libraries to point the mouse. in this program by showing small colour and with camera on computer it will detect and follow the color and simply i point that to windows mouse point so whenever i move my finger with color on it will follow it.